refactor: add handshake.go
